Simple message generators for the Twitter API
Move all the functions needed to generate a SoupMessage into a
single file. This should deal with all the crazy RESTful API
and their parameters, and return a simple SoupMessage with all
the bits already set, and ready to be queued/sent.

This API is not meant to be public, as of now, but it might be
exposed as a low level entry point for other implementations.