Huge overhauling of TwitterClient
Use the new Twitter API functions and remove the exposure of the
TwitterTimeline and TwitterUserList objects.

TwitterClient will emit a ::status-received and a ::user-received
each time a new Status and User object are received, respectively.
If we get a collection of statuses or users, we schedule a lazy
emission using a loop over an idle handler. This should keep the
library as much asynchronous as possible, and with as little
overhead as possible as well.

Warning: the reference counting hell is dangerously near; needs
testing.

TwitterClient still has gaping holes in it, but the tests we wrote
earlier work correctly, and it's possible to read and write statuses.