Stub out all the TwitterClient methods
The closure types need an overhaul to avoid copying always the same bits
all over the place. Instead of having different structures with the same
fields, use a parent structure and the great C structure alignment feature.

The methods are mostly stubbed out because of the missing message API,
but they can be implemented as soon as a Twitter API function is added.