Asynchronously load the user's profile image
When the user profile image is requested, check if it is already
present in the cache, using the URL SHA-1 checksum, and if it is,
asynchronously load that. Otherwise, use libsoup to retrieve it
from its URL, and cache the result locally.

At the end of each loading sequence emit the ::changed signal on
the user object and keep the GdkPixbuf around.