Return the timeline in ascending order
The timeline of statuses was returned in a newer to older order, but
this will force you to append to a list in case you want to display
the timeline in the same order. It also implied reversing the list
of statuses before returning it.

Instead, the timeline should return the statuses inside an older to
newer order, so that a model can just prepend every status. This
increases the responsivity of any UI that wants to emulate the
web interface, and present the statuses in a "newer first" sorting
order.