Add a string-to-timeval function
Twitter is using an insane variant of the CTime time format. This
not only destroys what little hope I had for humanity and instills
a much greater will of exacting furious and possibly highly painful
revenge upon web developers, but it also breaks SoupDate.

Hence, we need a custom function - like in the old days.