Implement the credentials verification
The verify_credentials user method was still not implemented. Since
we need it for checking whether the user name and password entered
by the user at her first login are valid, we need a working
twitter_client_verify_user().

The default implementation just sends the SoupMessage and checks
the message status, then emits the ::user-verified signal with a
boolean value, meaning we were verified.