Add the :url property to TwitterStatus
Each Twitter status has an URL associated to it, built from the
user screen name and the status id.

When we parse a TwitterStatus we should also build the URL and
expose it in the Status properties, so that we can use it to show
a link to the status, for instance.