Allow specifying a different base URL for the API
In order to use TwitterClient with another service that provides
a compatibility API with the default Twitter service we need to
decouple the Twitter API from the actual base URL.

This is done by adding a "base_url" parameter to all the API
wrapper functions inside twitter-api.[ch].

After that, we add a :base-url construct-only property to
TwitterClient, and we use the property to store the base URL
that we'll pass to the twitter_api_* wrappers.

Finally, in order to make construction simpler, we use an enum
to store the providers we know about, and a :provider property
to easily set them without using the URL, thus:

  g_object_new (TWITTER_TYPE_CLIENT,
                "provider", TWITTER_DEFAULT_PROVIDER,
                "email", email,
                "password", password,
                NULL);

is logically equivalent to:

  g_object_new (TWITTER_TYPE_CLIENT,
                "base-url", "http://twitter.com",
                "email", email,
                "password", password,
                NULL);

but without the pesky URL, or the usage of pre-processor macros.

Based on a patch by: Arnaud VALLAT <rno.rno@gmail.com>