Move User and Status definitions to their headers
Gtk-doc gets somewhat confused if we put every definition inside
a file with a different section. This leads to a wrong layout of
the API reference, and no cross-referencing for both TwitterUser
and TwitterStatus.

We leave the declarations inside the twitter-common.h header, as
we need them in a separate place to avoid circular dependency of
the Status and User headers.