Allow recoverable error handling for parsing user and status
When building a TwitterUser and a TwitterStatus out of their
JSON description we should have a way to programmatically
catch parse errors. Since an error of this kind is not a
programming error, it should be fully recoverable -- after all
we already allow recovering in case of network or service error.

This API change adds a GError parameter to the:

  twitter_status_load_from_data()
  twitter_user_load_from_data()

methods.

TwitterClient will relay a parse error to the ::status-received
and ::user-received signals.

The test unit for TwitterUser is also updated.