[client] Use handles for each requests
When calling TwitterClient methods it is difficult to associate
a signal emission to the request that actually resulted in the
signal being emitted.

In order to differentiate each request we can use an integer "handle";
each method resulting in the emission of a signal will also return
a "handle". The signal emitted will have the same handle of the
request as one of the parameters.