Rework the Twitter time-to-timestamp conversion
The conversion of the time string returned by Twitter objects to a
Unix timestamp is incorrect and unnecessarily convoluted.

When we parse the crazy string Twitter passes us we should always
get a UTC-relative struct tm out of strptime(); after that, we need
to call timegm() to get a "seconds from the Epoch" timestamp that
we can safely assign to a GTimeVal.

If the system does not provide timegm() then we can have a simple
fallback code path that does the same calculations.

This way, twitter_date_to_time_val() will return a UTC-relative
timestamp, exactly like Twitter.