[client] Expose Twitter's rate limiting
Twitter has rate limiting in place to avoid swamping their servers; for
this reason, with each request, it notifies the user about the maximum
amount of requests, and the remaining requests that can be performed.

This information should be exposed to clients using TwitterClient, so
that they can throttle the reloading of the user timeline to a reasonable
value that does not lead to Twitter thinking you are performing a DoS
on them.